requirements.txt
torch
transformers
deepspeed==0.15.1
tqdm
python-box
accelerate
Pillow
gymnasium
peft
wandb
opencv-python
shapely
termcolor
polyline
folium
geopy
numpy
easydict
pyyaml
requests
Pillow
openai
selenium==4.9.1
matplotlib
termcolor
opencv-python
torch
geopy
flask
flask_cors
flask_socketio
folium
polyline
prettytable
langchain
google-search-results
webdriver_manager
shapely
ipdb
pandas
gradio_client